new file:   app/Http/Controllers/CamaraController.php 	modified:   app/Http/Controllers/DeleteController.php 	new file:   app/Models/area.php 	modified:   app/Models/cvv_marca.php 	modified:   app/Models/cvv_modelo.php 	new file:   app/Models/sitio.php 	new file:   database/factories/areaFactory.php 	new file:   database/factories/sitioFactory.php 	new file:   database/migrations/2023_08_15_165640_areas.php 	new file:   database/migrations/2023_08_15_165650_sitios.php 	modified:   database/seeders/DatabaseSeeder.php 	modified:   public/js/camaras/function.js 	modified:   public/js/camaras/init.js 	modified:   resources/views/camaras.blade.php 	modified:   routes/web.php

+++ app/Http/Controllers/DeleteController.php
@@ -15,25 +15,32 @@ class DeleteController extends Controller
         $camara = cvv_camara::where('no_serie', $no_serie)->first();
     
         if ($camara) {
-            $fileName = strrchr($camara->foto_cam, '/');
 
-            $fileName = substr($fileName, 1);
+            if ($camara->foto_cam === null) {
+                // El campo "foto_cam" es NULL
+                // Realiza aquí la acción que deseas cuando la foto es NULL
+            } else {
+                $fileName = strrchr($camara->foto_cam, '/');
 
-            $file_path="public/imagenes_camaras/".$fileName;
-
-            if (Storage::exists($file_path)) {
-                if (Storage::delete($file_path)) {
-                    // Imagen eliminada exitosamente
+                $fileName = substr($fileName, 1);
+    
+                $file_path="public/imagenes_camaras/".$fileName;
+    
+                if (Storage::exists($file_path)) {
+                    if (Storage::delete($file_path)) {
+                        // Imagen eliminada exitosamente
+                    } else {
+                        return response()->json(['message' => 'Error al eliminar la imagen',
+                                                 'type'  => 'warning'
+                        ]);
+                    }
                 } else {
-                    return response()->json(['message' => 'Error al eliminar la imagen',
+                    return response()->json(['message' => 'La imagen no existe',
                                              'type'  => 'warning'
                     ]);
                 }
-            } else {
-                return response()->json(['message' => 'La imagen no existe',
-                                         'type'  => 'warning'
-                ]);
             }
+            
             $camara->delete();
             return response()->json(['type' => 'success', 
                                      'message' => 'Camara eliminada exitosamente',
